// alfcomponents/components/base/enum/AlfComponentType.enum.ts
export enum AlfComponentTypeEnum {
    Button = 'Button',
    Switch = 'Switch',
    Checkbox = 'Checkbox',
    RadioButton = 'RadioButton',
    Input = 'Input',
    Textarea = 'Textarea',
    Tabs = 'alf-tabs',
    Autocomplete = 'Autocomplete',
    Card = 'Card',
    CardTitle = 'CardTitle',
    CardBody = 'CardBody',
    CardActions = 'CardActions',
    CardHeader = 'CardHeader',
    CardFooter = 'CardFooter',
    Default = 'Default',
    CardImage = "CardImage",
    Paginator = "Paginator",
    SelectPanel = 'SelectPanel',
    SelectPanelOption = 'SelectPanelOption',
}